Editorial: Use internal spec value for param (#52)
Using an ECMAScript string value to represent internal state may cause
readers to incorrectly assume that the state is directly observable from
ECMAScript code. Replace with a internal specification value in
accordance with the latest practice in ECMA262 [1].

Modify the values so that they will be distinctive in the context of the
specification in order to facilitate tool-assisted search (e.g. via the
`grep` command-line utility or via "Ctrl+F" in a web browser).

[1] https://github.com/tc39/ecma262/pull/2566